// milo/ui/src/proto/turboci/data/test/v1/test_check_description_option.pb.ts
export interface TestCheckDescriptionOption {
  /**
   * A short title for the test check, to be used when displaying in places like the Chronicle UI
   * graph. Keep under 30 characters; if longer than 30 characters, may be truncated in UIs.
   */
  readonly title?:
    | string
    | undefined;
  /**
   * A message string with a more detailed description of the tests, with information on how to
   * format the text. Purely optional.
   */
  readonly displayMessage?: DisplayMessage | undefined;
}

function createBaseTestCheckDescriptionOption(): TestCheckDescriptionOption {
  return { title: undefined, displayMessage: undefined };
}

export const TestCheckDescriptionOption: MessageFns<TestCheckDescriptionOption> = {
  encode(message: TestCheckDescriptionOption, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.title !== undefined) {
      writer.uint32(10).string(message.title);
    }
    if (message.displayMessage !== undefined) {
      DisplayMessage.encode(message.displayMessage, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TestCheckDescriptionOption {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTestCheckDescriptionOption() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.title = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.displayMessage = DisplayMessage.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TestCheckDescriptionOption {
    return {
      title: isSet(object.title) ? globalThis.String(object.title) : undefined,
      displayMessage: isSet(object.displayMessage) ? DisplayMessage.fromJSON(object.displayMessage) : undefined,
    };
  },

  toJSON(message: TestCheckDescriptionOption): unknown {
    const obj: any = {};
    if (message.title !== undefined) {
      obj.title = message.title;
    }
    if (message.displayMessage !== undefined) {
      obj.displayMessage = DisplayMessage.toJSON(message.displayMessage);
    }
    return obj;
  },

  create(base?: DeepPartial<TestCheckDescriptionOption>): TestCheckDescriptionOption {
    return TestCheckDescriptionOption.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TestCheckDescriptionOption>): TestCheckDescriptionOption {
    const message = createBaseTestCheckDescriptionOption() as any;
    message.title = object.title ?? undefined;
    message.displayMessage = (object.displayMessage !== undefined && object.displayMessage !== null)
      ? DisplayMessage.fromPartial(object.displayMessage)
      : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}