// milo/ui/src/proto/turboci/data/test/v1/test_check_summary_result.pb.ts
export interface TestCheckSummaryResult {
  /**
   * Whether the test check succeeded or failed. "success = true" means that all tests succeeded.
   * Any failed tests result in "success = false". If the test failed, display_message will contain
   * human-readable information about the failure reasons. Additional product-specific failure
   * information should reside in a product-specific proto.
   */
  readonly success?:
    | boolean
    | undefined;
  /** A message string detailing the status of the tests with information on how to format the text. */
  readonly displayMessage?:
    | DisplayMessage
    | undefined;
  /** The total number of tests. */
  readonly testCount?:
    | number
    | undefined;
  /** The number of tests that succeeded. */
  readonly successCount?:
    | number
    | undefined;
  /** The number of tests that failed. */
  readonly failureCount?:
    | number
    | undefined;
  /** The number of tests that were skipped. */
  readonly skipCount?:
    | number
    | undefined;
  /**
   * A redirect URL to the external results store (ResultDB) that holds the detailed test results
   * for this test check.
   */
  readonly viewUrl?: string | undefined;
}

function createBaseTestCheckSummaryResult(): TestCheckSummaryResult {
  return {
    success: undefined,
    displayMessage: undefined,
    testCount: undefined,
    successCount: undefined,
    failureCount: undefined,
    skipCount: undefined,
    viewUrl: undefined,
  };
}

export const TestCheckSummaryResult: MessageFns<TestCheckSummaryResult> = {
  encode(message: TestCheckSummaryResult, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.success !== undefined) {
      writer.uint32(8).bool(message.success);
    }
    if (message.displayMessage !== undefined) {
      DisplayMessage.encode(message.displayMessage, writer.uint32(18).fork()).join();
    }
    if (message.testCount !== undefined) {
      writer.uint32(24).int32(message.testCount);
    }
    if (message.successCount !== undefined) {
      writer.uint32(32).int32(message.successCount);
    }
    if (message.failureCount !== undefined) {
      writer.uint32(40).int32(message.failureCount);
    }
    if (message.skipCount !== undefined) {
      writer.uint32(48).int32(message.skipCount);
    }
    if (message.viewUrl !== undefined) {
      writer.uint32(58).string(message.viewUrl);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TestCheckSummaryResult {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTestCheckSummaryResult() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.success = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.displayMessage = DisplayMessage.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.testCount = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.successCount = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.failureCount = reader.int32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.skipCount = reader.int32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.viewUrl = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TestCheckSummaryResult {
    return {
      success: isSet(object.success) ? globalThis.Boolean(object.success) : undefined,
      displayMessage: isSet(object.displayMessage) ? DisplayMessage.fromJSON(object.displayMessage) : undefined,
      testCount: isSet(object.testCount) ? globalThis.Number(object.testCount) : undefined,
      successCount: isSet(object.successCount) ? globalThis.Number(object.successCount) : undefined,
      failureCount: isSet(object.failureCount) ? globalThis.Number(object.failureCount) : undefined,
      skipCount: isSet(object.skipCount) ? globalThis.Number(object.skipCount) : undefined,
      viewUrl: isSet(object.viewUrl) ? globalThis.String(object.viewUrl) : undefined,
    };
  },

  toJSON(message: TestCheckSummaryResult): unknown {
    const obj: any = {};
    if (message.success !== undefined) {
      obj.success = message.success;
    }
    if (message.displayMessage !== undefined) {
      obj.displayMessage = DisplayMessage.toJSON(message.displayMessage);
    }
    if (message.testCount !== undefined) {
      obj.testCount = Math.round(message.testCount);
    }
    if (message.successCount !== undefined) {
      obj.successCount = Math.round(message.successCount);
    }
    if (message.failureCount !== undefined) {
      obj.failureCount = Math.round(message.failureCount);
    }
    if (message.skipCount !== undefined) {
      obj.skipCount = Math.round(message.skipCount);
    }
    if (message.viewUrl !== undefined) {
      obj.viewUrl = message.viewUrl;
    }
    return obj;
  },

  create(base?: DeepPartial<TestCheckSummaryResult>): TestCheckSummaryResult {
    return TestCheckSummaryResult.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TestCheckSummaryResult>): TestCheckSummaryResult {
    const message = createBaseTestCheckSummaryResult() as any;
    message.success = object.success ?? undefined;
    message.displayMessage = (object.displayMessage !== undefined && object.displayMessage !== null)
      ? DisplayMessage.fromPartial(object.displayMessage)
      : undefined;
    message.testCount = object.testCount ?? undefined;
    message.successCount = object.successCount ?? undefined;
    message.failureCount = object.failureCount ?? undefined;
    message.skipCount = object.skipCount ?? undefined;
    message.viewUrl = object.viewUrl ?? undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}